fix: remove duplicate directory
