--- trigger-with-condition/example/deploy/trigger.yaml
@@ -1,29 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: trigger
-  labels:
-    app: trigger
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: trigger
-  template:
-    metadata:
-      labels:
-        app: trigger
-      annotations:
-        dapr.io/enabled: "true"
-        dapr.io/app-id: "trigger"
-        dapr.io/log-as-json: "true"
-        dapr.io/app-protocol: "grpc"
-        dapr.io/app-port: "5050"
-    spec:
-      containers:
-        - name: trigger
-          image: trigger-with-condition:latest
-          imagePullPolicy: IfNotPresent
-          env:
-            - name: CONFIG
-              value: "eyJidXNDb21wb25lbnQiOiJ0cmlnZ2VyIiwiYnVzVG9waWMiOlt7Im5hbWUiOiJBIiwibmFtZXNwYWNlIjoiZGVmYXVsdCIsImV2ZW50U291cmNlIjoibXktZXZlbnRzb3VyY2UiLCJldmVudCI6InNhbXBsZS1vbmUifSx7Im5hbWUiOiJCIiwibmFtZXNwYWNlIjoiZGVmYXVsdCIsImV2ZW50U291cmNlIjoibXktZXZlbnRzb3VyY2UiLCJldmVudCI6InNhbXBsZS10d28ifV0sInN1YnNjcmliZXJzIjp7IkEgXHUwMDI2XHUwMDI2IEIiOnsidG9waWMiOiJtZXRyaWNzIn0sIkEgfHwgQiI6eyJzaW5rQ29tcG9uZW50IjoiaHR0cC1zaW5rIn19LCJwb3J0IjoiNTA1MCJ9"
