+++ trigger/example/deploy/events-producer.yaml
