--- trigger/example/deploy/trigger-producer.yaml
@@ -1,30 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: trigger-producer
-  labels:
-    app: triggerproducer
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: triggerproducer
-  template:
-    metadata:
-      labels:
-        app: triggerproducer
-      annotations:
-        dapr.io/enabled: "true"
-        dapr.io/app-id: "trigger-producer"
-        dapr.io/log-as-json: "true"
-        dapr.io/app-protocol: "grpc"
-    spec:
-      containers:
-        - name: producer
-          image: trigger-producer:latest
-          imagePullPolicy: IfNotPresent
-          env:
-            - name: TARGET_NAME
-              value: "nats-eventbus"
-            - name: TARGET_TOPIC
-              value: "default"
