--- trigger/example/deploy/trigger.yaml
@@ -1,29 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: trigger
-  labels:
-    app: trigger
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: trigger
-  template:
-    metadata:
-      labels:
-        app: trigger
-      annotations:
-        dapr.io/enabled: "true"
-        dapr.io/app-id: "trigger"
-        dapr.io/log-as-json: "true"
-        dapr.io/app-protocol: "grpc"
-        dapr.io/app-port: "5050"
-    spec:
-      containers:
-        - name: trigger
-          image: trigger:latest
-          imagePullPolicy: IfNotPresent
-          env:
-            - name: CONFIG
-              value: "eyJidXNDb21wb25lbnROYW1lIjoidHJpZ2dlciIsImJ1c1RvcGljIjoiZGVmYXVsdCIsInN1YnNjcmliZXJzIjpbeyJzaW5rQ29tcG9uZW50TmFtZSI6Imh0dHAtc2luayJ9LHsidG9waWNOYW1lIjoibWV0cmljcyJ9XSwicG9ydCI6IjUwNTAifQ=="
